[chore]: User dto 패키지 이동 (#139)
