+++ src/main/java/com/example/bookjourneybackend/domain/user/dto/request/PostUsersEmailRequest.java
@@ -1,4 +1,4 @@
-package com.example.bookjourneybackend.domain.user.domain.dto.request;
+package com.example.bookjourneybackend.domain.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -7,12 +7,10 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class PostUsersVerificationEmailRequest {
+public class PostUsersEmailRequest {
 
     @NotBlank(message = "이메일은 필수 입력 항목입니다.")
+    @Email(message = "올바른 이메일 형식이 아닙니다.")
     private String email;
-
-    @NotBlank(message = "인증할 코드를 입력해주세요.")
-    private String code;
 
 }
